backend/requirements.txt
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
fastapi==0.63.0
h11==0.12.0
pony==0.7.14
pycparser==2.20
pydantic==1.7.3
PyNaCl==1.4.0
six==1.15.0
starlette==0.13.6
uvicorn==0.13.3